完善添加组&修改组&删除组

+++ public/js/main.js
@@ -8,7 +8,10 @@ $(function () {
         bindBtn: function () {
             var self = this;
             $('#J_delete_all_btn').on('click', function () {
-                self.deleteHosts();
+                self.dialog.confirm('确定批量删除?', function () {
+                    self.deleteHosts();
+                });
+
             });
             $('#J_pause_all_btn').on('click', function () {
                 self.pauseHosts();
@@ -36,6 +39,97 @@ $(function () {
                 }
             });
 
+            //添加组
+            $('#J_add_group_btn').on('click', function () {
+                self.addGroup();
+            });
+
+            //删除组
+            $('.J_pane').on('click', '.J_delete_group', function () {
+                self.dialog.confirm('确定删除该组?', function () {
+                    self.deleteGroup();
+                });
+            });
+
+            //修改组名
+            $('.J_pane').on('click', '.J_edit_group', function () {
+                var name = $(this).parents('.J_pane').attr('data-label');
+                self.editGroup($(this).parents('.J_pane'), name);
+            });
+        },
+        /**
+         * 删除组
+         */
+        deleteGroup: function () {
+            $('#J_nav_tabs').find('li.active').remove();
+            $('#J_tab_content').find('.J_pane.active').remove();
+            $('#J_nav_tabs').find('li').eq(1).addClass('active');
+            $('#J_tab_content').find('.J_pane').eq(0).addClass('active');
+            this.startHosts();
+        },
+        /**
+         * 添加组
+         */
+        addGroup: function () {
+            if ($('.J_pane').size() > 8) {
+                this.dialog.alert('不能多于8个组哦.');
+                return false;
+            }
+            var self = this;
+            self.groupEditForm('添加组', '', function (groupName) {
+                $('.hosts .active').removeClass('active');
+                var key = new Date().getTime();
+                var tpl = '<li role="presentation" class="active"><a href="#_' + key + '" aria-controls="home" role="tab" data-toggle="tab">' + groupName + '</a></li>';
+                $('#J_nav_tabs').append(tpl);
+                var content_tpl = '<div role="tabpanel" data-label="' + groupName + '" class="J_pane tab-pane active" id="_' + key + '">\
+                    <ul class="hosts_items"><li class="item head">\
+                    <div class="allSelect">\
+                    <label><input type="checkbox" class="J_all_select"> 全选</label>\
+                    </div>\
+                    <div class="action">\
+                    <button class="btn btn-default btn-xs J_add_hosts" type="submit">添加Hosts</button>\
+                    <button class="btn btn-default btn-xs J_edit_group" type="submit">修改组名</button>\
+                    <button class="btn btn-default btn-xs J_delete_group" type="submit">删除组</button>\
+                    </div>\
+                    </li></ul></div>';
+                $('#J_tab_content').append(content_tpl);
+                self.startHosts();
+            });
+        },
+        /**
+         * 编辑组
+         */
+        editGroup: function (target, name) {
+            var self = this;
+            self.groupEditForm('修改组', name, function (groupName) {
+                target.attr('data-label', groupName);
+                $('#J_nav_tabs .active a').text(groupName);
+                self.startHosts();
+            });
+        },
+        /**
+         * 组编辑表单
+         * @param title
+         * @param groupName
+         * @param callback
+         */
+        groupEditForm: function (title, groupName, callback) {
+            var self = this;
+            var form = $('<form>\
+                <div class="form-group">\
+                <label for="J_newGroupName">组名</label>\
+            <input type="text" required class="form-control" name="groupName" value="' + groupName + '" id="J_newGroupName" placeholder="组名">\
+                </div>\
+            <button type="submit" class="btn btn-info">确定</button>\
+            <button type="button" class="btn btn-default" data-dismiss="modal">取消</button>\
+                </form>');
+            this.dialog.init(title, form);
+
+            form.on('submit', function () {
+                callback && callback($('#J_newGroupName').val().trim());
+                self.dialog.closeDialog();
+                return false;
+            });
         },
         /**
          * 绑定Tab
@@ -80,7 +174,7 @@ $(function () {
             var result = {};
             $('.J_pane').each(function () {
                 var items = [];
-                result[$(this).attr('data-lable')] = items;
+                result[$(this).attr('data-label')] = items;
                 $(this).find('input[name=item]').each(function () {
                     var _this = $(this);
                     var item = {};
@@ -177,7 +271,101 @@ $(function () {
                     callback && callback();
                 }
             });
+        },
+        /**
+         * dialog
+         */
+        dialog: {
+            /**
+             * 初始化Dialog
+             * @param title
+             * @param content
+             * @param sizeType
+             * @param autoRemove
+             */
+            init: function (title, content, sizeType, autoRemove) {
+                if (sizeType) {
+                    if (sizeType == 'sm') {
+                        $('#J_dialog_modal').removeClass('bs-example-modal-lg').addClass('bs-example-modal-sm');
+                        $('#J_dialog_type').removeClass('modal-lg').addClass('modal-sm');
+                    } else if (sizeType == null) {
+                        $('#J_dialog_modal').removeClass('bs-example-modal-lg');
+                        $('#J_dialog_type').removeClass('bs-example-modal-lg');
+                    } else {
+                        $('#J_dialog_modal').addClass('bs-example-modal-lg').removeClass('bs-example-modal-sm');
+                        $('#J_dialog_type').addClass('modal-lg').removeClass('modal-sm');
+                    }
+                }
+                if (autoRemove == true) {
+                    $('#J_dialog_modal').modal('show');
+                } else {
+                    $('#J_dialog_modal').modal({
+                        backdrop: 'static',
+                        keyboard: false
+                    });
+                }
+                $('#J_dialog_title').html(title);
+                if (typeof content == 'string') {
+                    $('#J_dialog_content').html(content);
+                } else {
+                    $('#J_dialog_content').empty().append(content);
+                }
+            },
+            /**
+             * alert
+             * @param msg
+             * @param callback
+             * @param sizeType
+             */
+            alert: function (msg, callback, sizeType) {
+                var self = this;
+                var tpl = '<div></div><div class="fn14"> ' + msg +
+                    '</div><div class="ac mt10 bt1 pt10">' +
+                    '<button type="button" id="J_confirm_btn" class="btn btn-success">确定</button></div></div>';
+                $('#J_dialog_modal').modal('show');
+                self.init('提示', tpl, sizeType || 'sm');
+                $('#J_confirm_btn').on('click', function () {
+                    self.closeDialog();
+                    callback && callback();
+                });
+            },
+            /**
+             * 确认
+             * @param msg
+             * @param confirmCallBack
+             */
+            confirm: function (msg, confirmCallBack) {
+                var self = this;
+                var tpl = '<div></div><div class="fn14"><i class="glyphicon glyphicon-question-sign orange"></i> ' + msg +
+                    '</div><div class="ar mt10 bt1 pt10">' +
+                    '<button type="button" id="J_confirm_btn" class="btn btn-success">确定</button>' +
+                    '<button type="button" id="J_cancel_btn" class="btn btn-default ml10">取消</button></div></div>';
+
+                $('#J_dialog_modal').modal('show');
+                self.init('请确认', tpl, 'sm');
+                $('#J_cancel_btn').on('click', self.closeDialog);
+
+                $('#J_confirm_btn').on('click', function () {
+                    self.closeDialog();
+                    setTimeout(function () {
+                        confirmCallBack();
+                    }, 500);
+                });
+            },
+            /**
+             * 关闭
+             * @param callback
+             */
+            closeDialog: function (callback) {
+                $('#J_dialog_modal').modal('hide');
+                if (typeof callback == 'function') {
+                    setTimeout(function () {
+                        callback();
+                    }, 600);
+                }
+            }
         }
+
     };
     _['init']();
 });
